chg: v3bwfile: Stop scaling with consensus weight
because when the observed bandwidth is higher than the consensus (for
example when the relay is new or was some time down), it's limited by
the previous consensus, not allowing it to grow.

Since the size of the data to download depends also on the consensus
weight, this results on lower measured bandwidth too.

Closes #40091.