new: scanner: do HTTP POST requests instead of GET
when the consensus parameter `bwscanner` is 2.

"Using upload instead of download allows the congestion control
algorithms to be under control of the sbws Tor client, instead of the
Exit relay. This means that sbws can know when `slow start` has exited."

This is the 1st optimization for sbws#40130.

The testing HTTP server has been replaced by the one implemented in
onbasca just cause a first try on implementing `do_POST` was
successful only for the 1st sbws request. It only implemnts POST RFC
2388.

Co-authored-by: Georg Koppen <gk@torproject.org>