fix: Remove asserts outside of tests
because they're removed anyway when using optimized mode and i don't
know why they were introduced.
Some of this code might need instead some checking or rasing
exceptions, but this version is in maintainance status.
In case this causes new errors, it can be detected by sbws hanging or
no progressing with measurements.
Now it's even possible to check the code with `bandit`.

Closes #40061.